# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Zawadi — Premium Kenyan Corporate Gifting Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /hampers:
    get:
      operationId: listHampers
      tags: [hampers]
      summary: List saved hamper configurations for a corporate
      parameters:
        - name: corporate_id
          in: query
          schema: { type: string }
      responses:
        "200":
          description: List of hampers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Hamper"
    post:
      operationId: createHamper
      tags: [hampers]
      summary: Create / save a custom hamper configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHamperInput"
      responses:
        "201":
          description: Created hamper
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hamper"
  /hampers/{id}:
    get:
      operationId: getHamper
      tags: [hampers]
      summary: Get a hamper with all items
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Hamper details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HamperWithItems"
    put:
      operationId: updateHamper
      tags: [hampers]
      summary: Update hamper items
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateHamperInput"
      responses:
        "200":
          description: Updated hamper
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hamper"
    delete:
      operationId: deleteHamper
      tags: [hampers]
      summary: Delete a hamper
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted

  # ── Corporates ────────────────────────────────────────────────────────────
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      parameters:
        - name: corporate_id
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: Paginated order list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get full order details
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Order with items and timeline
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderWithItems"
    put:
      operationId: updateOrder
      tags: [orders]
      summary: Update order (admin — status, delivery date, notes)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderInput"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/cancel:
    post:
      operationId: cancelOrder
      tags: [orders]
      summary: Cancel an order
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Cancelled order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # ── Recipients ────────────────────────────────────────────────────────────
...
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List invoices for a corporate
      parameters:
        - name: corporate_id
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
      responses:
        "200":
          description: List of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Generate a KRA-compliant invoice for an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceInput"
      responses:
        "201":
          description: Created invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get invoice details
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Invoice details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"

  # ── Dashboard ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Category:
      type: object
      required: [id, name, slug]
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        icon:
          type: string
          nullable: true
        product_count:
          type: integer

    # ── Supplier ──────────────────────────────────────────────────────────
    Supplier:
      type: object
      required: [id, name, county, is_verified, onboarding_status, created_at]
      properties:
        id:
          type: string
        name:
          type: string
        contact_name:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        county:
          type: string
        description:
          type: string
          nullable: true
        story:
          type: string
          nullable: true
        cover_image_url:
          type: string
          nullable: true
        logo_url:
          type: string
          nullable: true
        is_verified:
          type: boolean
        onboarding_status:
          type: string
          enum: [pending, under_review, approved, rejected]
        tags:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
...
    Product:
      type: object
      required: [id, name, category_id, supplier_id, unit_price, moq, lead_time_days, origin, is_active, created_at]
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        category_id:
          type: string
        category:
          $ref: "#/components/schemas/Category"
          nullable: true
        supplier_id:
          type: string
        supplier:
          $ref: "#/components/schemas/Supplier"
          nullable: true
        unit_price:
          type: number
          description: Price in KES per unit at MOQ
        bulk_tiers:
          type: array
          items:
            $ref: "#/components/schemas/BulkPriceTier"
        moq:
          type: integer
          description: Minimum order quantity
        lead_time_days:
          type: integer
        origin:
          type: string
          description: County/region of origin
        tags:
          type: array
          items:
            type: string
        occasion_tags:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            $ref: "#/components/schemas/ProductImage"
        is_active:
          type: boolean
        is_featured:
          type: boolean
        weight_grams:
          type: integer
          nullable: true
        dimensions_cm:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    BulkPriceTier:
      type: object
      required: [min_qty, price_per_unit]
      properties:
        min_qty:
          type: integer
        price_per_unit:
          type: number
    ProductImage:
      type: object
      required: [id, url, is_primary]
      properties:
        id:
          type: string
        url:
          type: string
        alt:
          type: string
          nullable: true
        is_primary:
          type: boolean
...
    HamperItem:
      type: object
      required: [id, product_id, quantity, unit_price]
      properties:
        id:
          type: string
        product_id:
          type: string
        product:
          $ref: "#/components/schemas/Product"
          nullable: true
        quantity:
          type: integer
        unit_price:
          type: number
    Hamper:
      type: object
      required: [id, name, total_price, status, created_at]
      properties:
        id:
          type: string
        corporate_id:
          type: string
          nullable: true
        name:
          type: string
        occasion:
          type: string
          nullable: true
        total_price:
          type: number
        item_count:
          type: integer
        status:
          type: string
          enum: [draft, saved, ordered]
        created_at:
          type: string
          format: date-time
    HamperWithItems:
      allOf:
        - $ref: "#/components/schemas/Hamper"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/HamperItem"
    CreateHamperInput:
      type: object
      required: [name, items]
      properties:
        corporate_id:
          type: string
          nullable: true
        name:
          type: string
        occasion:
          type: string
          nullable: true
        items:
          type: array
          items:
            type: object
            required: [product_id, quantity]
            properties:
              product_id:
                type: string
              quantity:
                type: integer
    UpdateHamperInput:
      type: object
      properties:
        name:
          type: string
        occasion:
          type: string
          nullable: true
        items:
          type: array
          items:
            type: object
            required: [product_id, quantity]
            properties:
              product_id:
                type: string
              quantity:
                type: integer

    # ── Corporate ─────────────────────────────────────────────────────────
...
    OrderItem:
      type: object
      required: [id, product_id, product_name, quantity, unit_price, line_total]
      properties:
        id:
          type: string
        product_id:
          type: string
        product_name:
          type: string
        quantity:
          type: integer
        unit_price:
          type: number
        line_total:
          type: number
        personalisation_text:
          type: string
          nullable: true
        branded_packaging:
          type: boolean
    Order:
      type: object
      required: [id, corporate_id, status, subtotal, vat, total, created_at]
      properties:
        id:
          type: string
        reference:
          type: string
          description: Human-readable order reference e.g. ZWD-2026-0042
        corporate_id:
          type: string
        corporate_name:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, confirmed, in_production, quality_check, packaging, dispatched, delivered, cancelled]
        subtotal:
          type: number
        vat:
          type: number
        total:
          type: number
        delivery_address:
          type: string
          nullable: true
        delivery_date:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
        recipient_count:
          type: integer
        created_at:
          type: string
          format: date-time
    OrderWithItems:
      allOf:
        - $ref: "#/components/schemas/Order"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/OrderItem"
            recipients:
              type: array
              items:
                $ref: "#/components/schemas/Recipient"
    OrderListResponse:
      type: object
      required: [items, total, limit, offset]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    CreateOrderInput:
      type: object
      required: [corporate_id, items, delivery_address]
      properties:
        corporate_id:
          type: string
        items:
          type: array
          items:
            type: object
            required: [product_id, quantity]
            properties:
              product_id:
                type: string
              quantity:
                type: integer
              branded_packaging:
                type: boolean
              personalisation_text:
                type: string
                nullable: true
        delivery_address:
          type: string
        delivery_date:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
        quote_id:
          type: string
          nullable: true
    UpdateOrderInput:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, in_production, quality_check, packaging, dispatched, delivered, cancelled]
        delivery_date:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true

    # ── Recipient ─────────────────────────────────────────────────────────
    Recipient:
      type: object
      required: [id, order_id, name]
      properties:
        id:
          type: string
        order_id:
          type: string
        name:
          type: string
        email:
          type: string
          nullable: true
        department:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        personalisation_message:
          type: string
          nullable: true
        delivery_address:
          type: string
          nullable: true
...
    Invoice:
      type: object
      required: [id, order_id, invoice_number, amount, status, created_at]
      properties:
        id:
          type: string
        order_id:
          type: string
        corporate_id:
          type: string
          nullable: true
        invoice_number:
          type: string
        kra_pin:
          type: string
          nullable: true
        amount:
          type: number
        vat_amount:
          type: number
        total_amount:
          type: number
        status:
          type: string
          enum: [draft, sent, paid, overdue, cancelled]
        due_date:
          type: string
          format: date-time
          nullable: true
        paid_at:
          type: string
          format: date-time
          nullable: true
        pdf_url:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    CreateInvoiceInput:
      type: object
      required: [order_id]
      properties:
        order_id:
          type: string
        due_date:
          type: string
          format: date-time
          nullable: true

    # ── Dashboard ─────────────────────────────────────────────────────────
...